Fixed timing issue in Launchpad sample.
Added the correct frequency of the submaster clock in the Launchpad's
system definition file. The clock is actually sourced from a 32 KHz
external oscillator whose frequency is increased by a factor of 32 for a
final frequency of around 1 MHz. Also fixed the maximum timer value,
which was to large by 1, leading to an overflow error in the compare
value calculations.